Sucrose as Organ?
Sucrose may be not be considered as an organ due to the fact that it is
a component of the internode. The substrate pool of entire plant should
be partitioned among internodes and be considered as a total substrate
pool or total sugars within each internode. Then, the process of
fractioning total sugars in each internode to sucrose/hexose. This
process could be correlated to internode age, weight, RWC, total sugars.

see:
Modelling genetic and environmental control of biomass partitioning at
plant and phytomer level of sugarcane grown in controlled environments
(Singels & Inman-Bamber, 2011)

Sugarcane Internode Composition During Crop Development (Lingle &
Thomson, 2012)