RWF - WEIBUL (Debugged)
TRWUP is too low due to the approach used (swcon1,2,3 - CERES).

Looking on CANEGRO outputs, its possible to find RLD around 9 cm cm-3,
this could be a compensation for CERES trwup is too low. On literature,
typical values of RLD for sugarcane are below 1 cm cm-3 (Laclau & Laclau
2009) or 25 cm cm-2.

trwup is too low because CERES method restric RWU when LL are high
and/or diference between (DUL - LL)  are small (this is what happening
to our soil on Areão! mean LL = 0.25; mean (DUL-LL) = 0.08)